+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change log for CombineAutopilotHashes
 
 
+## 1.0.1 -December 1, 2021
+
+- Added addition functions to check path; incase script is ran in ISE
+- Removed CombineHashes.csv if found; script would append to current list otherwise
+- Added "Group Tag" support; Enforce Header selection even if does not exist in some csv
+- Added Output to see status
+
 ## 1.0.0 - November 30, 2021
 
 - initial build